Interim update

+++ firmware/rle_test.h
@@ -7,8 +7,8 @@
 #include "util.h"
 
 void rle_test(void);
-void prepare_sample(int *samp_data);
-void prepare_byte(uint8_t *data, int *sample_data);
+void prepare_sample(uint64_t *samp_data);
+void prepare_byte(uint8_t *data, uint64_t *sample_data);
 void debug_print_rle_compressed_blocks(bitstream *compressed_b_stream, uint8_t sig);
 void uart_print_compressed(bitstream* b_streams);
